requirements.txt
apispec==6.6.1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
flask-smorest==0.44.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.21.3
numpy==2.0.0
packaging==24.1
pandas==2.2.2
py-dotenv==0.1
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tzdata==2024.1
webargs==8.4.0
Werkzeug==3.0.3
